Restore special case for trivial maps (bug #1837422)
A trivial map means a map having only a single subsector, and no nodes.

There was a special case for this in R_PointInSubsector in the original
source release, but it was removed from Boom. (Several other places in
the code rely on the fact that (-1)&NF_SUBSECTOR is non-zero)

Then allow these maps to be loaded in P_LoadNodes, but inform the user.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@2889 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e