P_DrawPSprite: Add fudge factor for weapon brightness
An old complaint about PrBoom is that it draws player weapons too dark.
Long ago, the lightlevel to colourmap lookup was unified into a single
function, R_ColourMap. Possibly because psprites are drawn unscaled onto
the player view, the scale factor passed to R_ColourMap by R_DrawPSprite
was that of an unscaled mobj sprite. However this means it was being
drawn with the colourmap of a mobj standing 160 units away, which is a
lot darker than one as close as your weapon is.

We can fix this by introducing a fudge factor to the colourmap lookup in
R_DrawPSprite. The one chosen here was found after some experimentation
to have the smallest difference from PrBoom-plus. Therefore I believe
this to be a good compromise solution, rather than simply reverting back
to Doom's messy lookup tables.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@2906 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e