Rewrite logic of intermission screen time counters
When a PWAD is loaded and if intermission screen counting is not
skipped, pistol sounds are still played even after the time and total
time counters have finished counting up. This is because the code is
still looping over the (suppressed) par time.

Instead, independently update each time counter. If they're all
finished, play a barrel explosion and move onto the next state;
otherwise, play gunshots in a loop as usual.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@2980 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e