Always use the C version of D_abs
Prefer the inline function version over the macro, but fix its
declaration: it both takes as a parameter, and returns, a fixed_t.
Also, use CONSTFUNC instead of const.

[ In case anyone is concerned about the "minor performance penalty"
mentioned in the previous commit: on my system a timedemo of 30ns7155
(162515 gametics) went from 35579 realtics to 35786 - an increase of
207, or about 0.6%. FPS dropped from 159.9 to 158.9. In short, I believe
an extra six seconds in seventeen minutes is no cause for concern. ]

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@3203 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e