P_GetAmmoLevel: return percentage ammo level of weapon
New function, derived from P_CheckAmmo and the HUD and status bar code.
Returns int from 0-100 indicating ammunition level of the given weapon.

0/100 means the weapon is empty/full respectively. It takes care to only
return 0 if the weapon cannot be fired, handling the SSG/BFG correctly.

Purpose is to make the weapon code itself, and the HUD/status bar, use
exactly the same code to determine whether a weapon is empty or not.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@3637 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e