Merge r3672 from plus (restore parameterless -warp)
The function GetFirstMap is present in Boom, but missing in MBF. It
appears that when MBF was copied into LxDoom, the call to GetFirstMap in
D_DoomMain was lost, and the function itself was left behind, orphaned.

Anyway now you can do -warp without a parameter (or "-warp *") to jump
to the first map in a PWAD even if it's not E1M1/MAP01. Useful!

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@3680 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e