p_floor.c: fix bitwise/boolean warning in EV_DoDonut
This expression, present in linuxdoom-1.10, gives a warning in gcc-4.4:

    !s2->lines[i]->flags & ML_TWOSIDED

Due to operator precedence it is equivalent to:

    (!s2->lines[i]->flags) & ML_TWOSIDED

which, unfortunately, is always zero, since ML_TWOSIDED is 4. Put in the
brackets to silence the compiler. The correct expression is, of course

    !(s2->lines[i]->flags & ML_TWOSIDED)

but whether it can be unconditionally fixed is an open question.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@3735 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e