d_deh.c: apply fraggle's patch for multiple string replacements
If two DeHackEd patches try to replace the same string, the second will
fail as the Boom DeHackEd parser would try to match against the string
it read from the first patch instead of the original. This patch fixes
it by keeping a pointer to the original when a string is modified.

Slightly reformatted from the version posted at
<http://www.doomworld.com/vb/post/978852>. See also plus r3975

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@3976 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e