rdatawad: import data wad builder from rboom
This is a reimplementation of just enough of deutex to be able to
build the data wad. It handles:

- raw lumps (inserting a file as a lump without modification)
- sounds (converting wav files of a very specific format)
- graphics (applying a palette to a raw image in ppm format and
  converting it into Doom's patch format)
- sprites (as graphics, but specifying an insertion point)

It is very minimal and not at all user friendly, but as the only
intended user is data/Makefile.am, I do not see this as a problem.

Motivation: I tried to preprocess the deutex directives file to omit
non-free lumps, but deutex did not work well with a combination of a
modified directives file in the build directory, and the resources in
the source directory. So I wrote something controlled from the command
line with explicitly-specified file search paths.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@4148 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e