Avoid use of PATH_MAX
Get storage from malloc instead, calling snprintf twice to get the sizes
needed in most cases.

Also, bug fix: in FindResponseFile, when looking in I_DoomExeDir for the
file, it was not correctly path-separated from the directory. Previous
code merely appended it, which is not correct, as I_DoomExeDir's return
value does not end in a slash. That this went undiscovered for so long
shows how much response files are actually used, I think. Note you can
achieve the same effect of a parameter @rsp with $(<rsp) in the shell.

git-svn-id: https://svn.prboom.org/repos/trunk/prboom2@4235 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e